wip2

+++ packages/onchain/contracts/balancer/IBPool.sol
@@ -44,14 +44,12 @@ interface IBPool {
     uint256 maxPrice
   ) external returns (uint256 tokenAmountOut, uint256 spotPriceAfter);
 
-     function calcInGivenOut(
-        uint tokenBalanceIn,
-        uint tokenWeightIn,
-        uint tokenBalanceOut,
-        uint tokenWeightOut,
-        uint tokenAmountOut,
-        uint swapFee
-    )
-        external pure
-        returns (uint tokenAmountIn);
+  function calcInGivenOut(
+    uint256 tokenBalanceIn,
+    uint256 tokenWeightIn,
+    uint256 tokenBalanceOut,
+    uint256 tokenWeightOut,
+    uint256 tokenAmountOut,
+    uint256 swapFee
+  ) external pure returns (uint256 tokenAmountIn);
 }
